// resources/js/modules/market/redux/market.types.ts
enum MarketTypes {
    MARKET_ACTION_START = "MARKET_ACTION_START",
    MARKET_GET_SUCCESS = "MARKET_GET_SUCCESS",
    MARKET_GET_FAILED = "MARKET_GET_FAILED",
    MARKET_CREATE_SUCCESS = "MARKET_CREATE_SUCCESS",
    MARKET_CREATE_FAILED = "MARKET_CREATE_FAILED",
    MARKET_GET_BY_ID_SUCCESS = "MARKET_GET_BY_ID_SUCCESS",
    MARKET_GET_BY_ID_FAILED = "MARKET_GET_BY_ID_FAILED",
    MARKET_UPDATE_SUCCESS = "MARKET_UPDATE_SUCCESS",
    MARKET_UPDATE_FAILED = "MARKET_UPDATE_FAILED",
    MARKET_DELETE_SUCCESS = "MARKET_DELETE_SUCCESS",
    MARKET_DELETE_FAILED = "MARKET_DELETE_FAILED",
    MARKET_ACTION_END = "MARKET_ACTION_END",
}

export default MarketTypes;